requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
azure-search-documents
azure-ai-documentintelligence
azure-identity
azure-storage-blob
chromadb
langchain-chroma
pypdf
tiktoken
python-dotenv
jupyter
ipykernel
deepeval
anthropic
langgraph
sqlalchemy
